+++ examples/github-repository-webhook/versions.tf
@@ -2,7 +2,6 @@ terraform {
   required_version = ">= 0.12.7"
 
   required_providers {
-    aws    = ">= 2.68"
-    random = ">= 2.0"
+    aws = ">= 2.68"
   }
 }